requirements.txt
apache-airflow>=2.7.0
requests>=2.31.0
pandas>=2.1.0
mysql-connector-python>=8.1.0 